Popravljene veze i dodani novi entiteti. Ispravljene sitne greške.

+++ doctor/src/main/java/com/sbnz/doctor/controller/IngredientController.java
@@ -2,6 +2,8 @@ package com.sbnz.doctor.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,6 +18,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.sbnz.doctor.dto.IngredientDTO;
+import com.sbnz.doctor.dto.UserDTO;
 import com.sbnz.doctor.interfaces.services.IngredientServiceInterface;
 
 /**
@@ -25,7 +28,7 @@ import com.sbnz.doctor.interfaces.services.IngredientServiceInterface;
 @RestController
 @RequestMapping(value = "/ingredient")
 public class IngredientController {
-	
+
 	@Autowired
 	private IngredientServiceInterface service;
 
@@ -53,7 +56,18 @@ public class IngredientController {
 	}
 
 	@RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON, produces = MediaType.APPLICATION_JSON)
-	public ResponseEntity<IngredientDTO> add(@Validated @RequestBody IngredientDTO body, Errors errors) {
+	public ResponseEntity<IngredientDTO> add(@Validated @RequestBody IngredientDTO body, Errors errors,
+			@Context HttpServletRequest request) {
+		UserDTO user = (UserDTO) request.getSession().getAttribute("user");
+
+		if (user == null) {
+			return new ResponseEntity<IngredientDTO>(body, HttpStatus.BAD_REQUEST);
+		}
+
+		if (user.getUserType() != 'A') {
+			return new ResponseEntity<IngredientDTO>(body, HttpStatus.FORBIDDEN);
+		}
+
 		if (errors.hasErrors()) {
 			return new ResponseEntity<IngredientDTO>(body, HttpStatus.BAD_REQUEST);
 		}
@@ -68,7 +82,18 @@ public class IngredientController {
 	}
 
 	@RequestMapping(method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON, produces = MediaType.APPLICATION_JSON)
-	public ResponseEntity<IngredientDTO> edit(@Validated @RequestBody IngredientDTO dto, Errors errors) {
+	public ResponseEntity<IngredientDTO> edit(@Validated @RequestBody IngredientDTO dto, Errors errors,
+			@Context HttpServletRequest request) {
+		UserDTO user = (UserDTO) request.getSession().getAttribute("user");
+
+		if (user == null) {
+			return new ResponseEntity<IngredientDTO>(dto, HttpStatus.BAD_REQUEST);
+		}
+
+		if (user.getUserType() != 'A') {
+			return new ResponseEntity<IngredientDTO>(dto, HttpStatus.FORBIDDEN);
+		}
+
 		if (errors.hasErrors()) {
 			return new ResponseEntity<IngredientDTO>(dto, HttpStatus.BAD_REQUEST);
 		}
@@ -81,7 +106,17 @@ public class IngredientController {
 	}
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON)
-	public ResponseEntity<IngredientDTO> delete(@PathVariable int id) {
+	public ResponseEntity<IngredientDTO> delete(@PathVariable int id, @Context HttpServletRequest request) {
+		UserDTO user = (UserDTO) request.getSession().getAttribute("user");
+
+		if (user == null) {
+			return new ResponseEntity<IngredientDTO>(HttpStatus.BAD_REQUEST);
+		}
+
+		if (user.getUserType() != 'A') {
+			return new ResponseEntity<IngredientDTO>(HttpStatus.FORBIDDEN);
+		}
+
 		IngredientDTO dto = service.Delete(id);
 		if (dto == null) {
 			new ResponseEntity<IngredientDTO>(dto, HttpStatus.NOT_FOUND);
